FizzBuzz: Move FizzBuzz to a separate file
Extract integration to its own file to separate functionality from testing behavior
