+++ clean-code-katas/function-katas/tests/fizzbuzz.tests/FizzBuzzTests.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using FluentAssertions;
+﻿using FluentAssertions;
 using Xunit;
 
 namespace fizzbuzz.tests
@@ -20,31 +19,6 @@ namespace fizzbuzz.tests
             _target.DoFizzBuzz();
 
             expected.Should().BeEquivalentTo(_target.Output);
-        }
-    }
-
-    internal class FizzBuzz
-    {
-        private readonly NumberGenerator _numberGenerator;
-        private readonly FizzBuzzConverter _fizzBuzzConverter;
-        private readonly TextDisplay _textDisplay;
-        private readonly Action<string> _displayAction;
-        public string Output { get; set; }
-
-        public FizzBuzz()
-        {
-            _numberGenerator = new NumberGenerator();
-            _fizzBuzzConverter = new FizzBuzzConverter();
-            _textDisplay = new TextDisplay();
-            _displayAction = (t) => Output = t;
-
-        }
-
-        public void DoFizzBuzz()
-        {
-            var numbers = _numberGenerator.GenerateNumbers();
-            var fizzBuzzTexts = _fizzBuzzConverter.ConvertToFizzBuzzTexts(numbers);
-            _textDisplay.DisplayTexts(fizzBuzzTexts, _displayAction);
         }
     }
 }
